Grid_Compress;Shift_Opposite;Grid_Merge--> created
Shift_Opposite function:
when we press the arrow buttons all the numbers will shift towards the arrow side so here in this function we are 2 loops so that we can iterate in a 2d gird for example (0,0) is 2 when we give the command as right the digit 2 should change to position (0,3)
Grid_Compress:
In this function, we are checking for the non zero elements, and using a temp variable we are updating the gridCell variable which we created with all zero initially 
Grid_Merge:
This function deals with the multiplying the elements with the same value in the gridCell. So here Our logic is Grid_Compress gives the elements after shifting and this grid merge will multiply the number by 2 if they are equal else they leave the same as shifted.